+++ app/src/main/java/com/alilopez/kt_demohilt/features/exercise/domain/repositories/ExerciseRepository.kt
@@ -5,4 +5,5 @@ import com.alilopez.kt_demohilt.features.exercise.domain.entities.Exercise
 interface ExerciseRepository {
     suspend fun getExercises(): List<Exercise>
     suspend fun getExercisesByBodyPart(bodyPart: String): List<Exercise>
+    suspend fun getLocalExercises(): List<Exercise>
 }