+++ app/src/main/java/com/alilopez/kt_demohilt/features/exercise/domain/usecases/GetLocalExercisesUseCase.kt
@@ -1,0 +1,20 @@
+package com.alilopez.kt_demohilt.features.exercise.domain.usecases
+
+import com.alilopez.kt_demohilt.features.exercise.domain.entities.Exercise
+import com.alilopez.kt_demohilt.features.exercise.domain.repositories.ExerciseRepository
+import javax.inject.Inject
+
+class GetLocalExercisesUseCase @Inject constructor(
+    private val repository: ExerciseRepository
+) {
+
+    suspend operator fun invoke(): Result<List<Exercise>> {
+        return try {
+            val exercises = repository.getLocalExercises()
+            Result.success(exercises)
+        } catch (e: Exception) {
+            Result.failure(e)
+        }
+    }
+}
+